# web/openapi/openapi.json
{
  "components": {
    "schemas": {
      "Body_create_document_api_v1_documents_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "title": "Body_create_document_api_v1_documents_post",
        "type": "object"
      },
      "Body_replace_document_api_v1_documents__source_id__put": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "title": "Body_replace_document_api_v1_documents__source_id__put",
        "type": "object"
      },
      "CatalogDocumentResponse": {
        "description": "Wire projection of one uploaded catalog row (sanitized diagnostics).",
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "content_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Format"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "has_error": {
            "title": "Has Error",
            "type": "boolean"
          },
          "has_stored_content": {
            "default": true,
            "title": "Has Stored Content",
            "type": "boolean"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "enum": ["pending", "ready", "failed", "degraded"],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "uploaded_at": {
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "source_type",
          "file_name",
          "status",
          "uploaded_at",
          "chunk_count",
          "has_error"
        ],
        "title": "CatalogDocumentResponse",
        "type": "object"
      },
...
      "DocumentChunkListResponse": {
        "description": "Stored chunks for one catalogued document.",
        "properties": {
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/DocumentChunkResponse"
            },
            "title": "Chunks",
            "type": "array"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          }
        },
        "required": ["chunks", "has_more"],
        "title": "DocumentChunkListResponse",
        "type": "object"
      },
      "DocumentChunkResponse": {
        "description": "Wire projection of one stored chunk for the documents UI.\n\nScalar fields are an explicit allowlist. ``extra`` forwards connector\nmetadata from ``SourceMetadata.extra`` without a key filter \u2014 callers must\nnot treat it as a closed schema.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "content_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Format"
          },
          "extra": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Extra",
            "type": "object"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": ["index", "content", "source_id", "source_type", "extra"],
        "title": "DocumentChunkResponse",
        "type": "object"
      },
      "DocumentListResponse": {
        "description": "Uploaded-document catalog for the documents UI.",
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/CatalogDocumentResponse"
            },
            "title": "Documents",
            "type": "array"
          }
        },
        "required": ["documents"],
        "title": "DocumentListResponse",
        "type": "object"
      },
...
      "HubSourceType": {
        "description": "Hub catalog kinds accepted by chunk-inspect query params.\n\nMirrors ``domain.knowledge.HUB_SOURCE_TYPES`` as a named OpenAPI schema so\nthe wire contract stays documented and does not silently widen with\n``SourceType``.",
        "enum": ["knowledge_document", "google_drive", "github"],
        "title": "HubSourceType",
        "type": "string"
      },
...
      "Problem": {
        "description": "RFC 9457 Problem Details plus Kernector extensions.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": ["type", "title", "status", "detail", "code"],
        "title": "Problem",
        "type": "object"
      },
      "ProblemError": {
        "description": "Field-level Problem Details extension entry.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "pointer": {
            "title": "Pointer",
            "type": "string"
          }
        },
        "required": ["pointer", "detail"],
        "title": "ProblemError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Minimal versioned HTTP adapter over composition. Unknown paths return RFC 9457 Problem Details with status 404 (app-level; not declared per operation).",
    "title": "Kernector API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "description": "Return upload and Google Drive catalog rows for the documents UI.",
        "operationId": "list_documents_api_v1_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "List Documents",
        "tags": ["documents"]
      },
      "post": {
        "description": "Upload a new document; always allocates a system-managed source ID.",
        "operationId": "create_document_api_v1_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_document_api_v1_documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogDocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Payload too large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Create Document",
        "tags": ["documents"]
      }
    },
    "/api/v1/documents/{source_id}": {
      "delete": {
        "description": "Delete chunks and catalog row. Unknown IDs are a deliberate 204 no-op.\n\n``source_id`` uses a path converter so GitHub ids that contain ``/``\n(``owner/repo:path/to/file``) still route after percent-decoding.",
        "operationId": "delete_document_api_v1_documents__source_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Delete Document",
        "tags": ["documents"]
      },
      "put": {
        "description": "Replace document content under the same source ID.",
        "operationId": "replace_document_api_v1_documents__source_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_replace_document_api_v1_documents__source_id__put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogDocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Payload too large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Replace Document",
        "tags": ["documents"]
      }
    },
    "/api/v1/documents/{source_id}/chunks": {
      "get": {
        "description": "Return a page of stored chunks for one catalogued source.",
        "operationId": "list_document_chunks_api_v1_documents__source_id__chunks_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/HubSourceType"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentChunkListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "List Document Chunks",
        "tags": ["documents"]
      }
    },
    "/api/v1/documents/{source_id}/content": {
      "get": {
        "description": "Return original uploaded bytes for inline display.",
        "operationId": "get_document_content_api_v1_documents__source_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "text/plain; charset=utf-8": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Original document content"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Get Document Content",
        "tags": ["documents"]
      }
    },
    "/api/v1/documents/{source_id}/download": {
      "get": {
        "description": "Return original uploaded bytes as an attachment.",
        "operationId": "download_document_api_v1_documents__source_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "text/plain; charset=utf-8": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Original document download"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Download Document",
        "tags": ["documents"]
      }
    },
...
  }
}